Changed: Relocate builder tests and add MAX_SEED constant
Moved builder-specific tests from ModelCatalog to ModelCatalogBuilder
where they belong, and extracted the hardcoded seed limit into a named
constant for clarity.

Changes:
- Moved 5 builder tests from mod.rs to builder.rs (collisions, reset, seed exhaustion, token limits)
- Added MAX_SEED constant (u8::MAX = 255) for hash reseed limit
- Replaced all u8::MAX usages with MAX_SEED constant

Benefits:
- Tests are now colocated with the code they test
- Hardcoded value has a meaningful name with documentation
- Makes the seed limit more explicit and discoverable
